// swiyu-ecosystem-portal-ui/src/app/api/generated/model/page.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {RegistrationResponse} from './registration-response';

/**
 * Representing a Page used for pagination
 */
export interface Page {
  content: Array<RegistrationResponse>;
  totalPages: number;
  totalElements: number;
  size: number;
  number: number;
}